requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
httpx==0.27.2
python-jose==3.3.0
pydantic-settings==2.6.1
PyJWT==2.9.0
